requirements.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.4
action-tutorials-py==0.33.4
actionlib-msgs==5.3.5
ament-cmake-test==2.5.2
ament-copyright==0.17.1
ament-cppcheck==0.17.1
ament-cpplint==0.17.1
ament-flake8==0.17.1
ament-index-python==1.8.1
ament-lint==0.17.1
ament-lint-cmake==0.17.1
ament-package==0.16.3
ament-pep257==0.17.1
ament-uncrustify==0.17.1
ament-xmllint==0.17.1
angles==1.16.0
builtin-interfaces==2.0.2
catkin-pkg==1.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
composition-interfaces==2.0.2
contourpy==1.2.1
cv-bridge==4.1.0
cycler==0.12.1
demo-nodes-py==0.33.4
diagnostic-msgs==5.3.5
diagnostic-updater==3.1.2
docutils==0.21.2
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.4
examples-rclpy-minimal-action-client==0.19.4
examples-rclpy-minimal-action-server==0.19.4
examples-rclpy-minimal-client==0.19.4
examples-rclpy-minimal-publisher==0.19.4
examples-rclpy-minimal-service==0.19.4
examples-rclpy-minimal-subscriber==0.19.4
fake-useragent==1.5.1
filelock==3.15.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
geometry-msgs==5.3.5
humanfriendly==10.0
idna==3.7
image-geometry==4.1.0
imread_from_url==0.1.3
imutils==0.5.4
interactive-markers==2.5.4
Jinja2==3.1.4
kiwisolver==1.4.5
laser-geometry==2.7.0
launch==3.4.2
launch-ros==0.26.5
launch-testing==3.4.2
launch-testing-ros==0.26.5
launch-xml==3.4.2
launch-yaml==3.4.2
-e git+https://github.com/jinzhengguang/RPEA.git@d036f3f8afdd155f9261edb93b61c64c90851387#egg=lidar_det
lifecycle-msgs==2.0.2
logging-demo==0.33.4
map-msgs==2.4.1
MarkupSafe==2.1.5
matplotlib==3.9.0
message-filters==4.11.1
mpmath==1.3.0
nav-msgs==5.3.5
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
osrf-pycommon==2.1.4
packaging==24.1
pandas==2.2.2
pcl-msgs==1.0.0
pendulum-msgs==0.33.4
pillow==10.3.0
protobuf==5.27.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-qt-binding==2.2.1
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==2.7.4
qt-gui==2.7.4
qt-gui-cpp==2.7.4
qt-gui-py-common==2.7.4
quality-of-service-demo-py==0.33.4
rcl-interfaces==2.0.2
rclpy==7.1.1
rcutils==6.7.1
requests==2.32.3
resource-retriever==3.4.3
rmw-dds-common==3.1.0
ros2_numpy @ file:///home/wicomai/ros2_numpy
ros2action==0.32.1
ros2bag==0.26.4
ros2bag-mcap-cli==0.26.4
ros2bag-sqlite3-cli==0.26.4
ros2cli==0.32.1
ros2component==0.32.1
ros2doctor==0.32.1
ros2interface==0.32.1
ros2launch==0.26.5
ros2lifecycle==0.32.1
ros2multicast==0.32.1
ros2node==0.32.1
ros2param==0.32.1
ros2pkg==0.32.1
ros2run==0.32.1
ros2service==0.32.1
ros2topic==0.32.1
rosbag2-interfaces==0.26.4
rosbag2-py==0.26.4
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.3
rosidl-cli==4.6.3
rosidl-cmake==4.6.3
rosidl-generator-c==4.6.3
rosidl-generator-cpp==4.6.3
rosidl-generator-py==0.22.0
rosidl-generator-type-description==4.6.3
rosidl-parser==4.6.3
rosidl-pycommon==4.6.3
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.0
rosidl-typesupport-fastrtps-cpp==3.6.0
rosidl-typesupport-introspection-c==4.6.3
rosidl-typesupport-introspection-cpp==4.6.3
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.3
rqt-bag-plugins==1.5.3
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.0
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.2
scipy==1.13.1
seaborn==0.13.2
sensor-msgs==5.3.5
sensor-msgs-py==5.3.5
service-msgs==2.0.2
setuptools==70.3.0
shape-msgs==5.3.5
six==1.16.0
sros2==0.13.2
statistics-msgs==2.0.2
std-msgs==5.3.5
std-srvs==5.3.5
stereo-msgs==5.3.5
sympy==1.12.1
teleop-twist-keyboard==2.4.0
tf-transformations==1.0.0
tf2-geometry-msgs==0.36.4
tf2-kdl==0.36.4
tf2-msgs==0.36.4
tf2-py==0.36.4
tf2-ros-py==0.36.4
tf2-sensor-msgs==0.36.4
tf2-tools==0.36.4
topic-monitor==0.33.4
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trajectory-msgs==5.3.5
turtlesim==1.8.3
type-description-interfaces==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.38
ultralytics-thop==2.0.0
unique-identifier-msgs==2.5.0
urllib3==2.2.2
velodyne-msgs==2.3.0
visualization-msgs==5.3.5
